+++ exporters/stats/stackdriver/src/main/java/io/opencensus/exporter/stats/stackdriver/StackdriverStatsExporter.java
@@ -79,11 +79,17 @@ public final class StackdriverStatsExporter {
       MetricServiceClient metricServiceClient,
       Duration exportInterval,
       ViewManager viewManager,
-      MonitoredResource monitoredResource) {
+      MonitoredResource monitoredResource,
+      @Nullable String metricNamePrefix) {
     checkArgument(exportInterval.compareTo(ZERO) > 0, "Duration must be positive");
     StackdriverExporterWorker worker =
         new StackdriverExporterWorker(
-            projectId, metricServiceClient, exportInterval, viewManager, monitoredResource);
+            projectId,
+            metricServiceClient,
+            exportInterval,
+            viewManager,
+            monitoredResource,
+            metricNamePrefix);
     this.workerThread = new DaemonThreadFactory().newThread(worker);
   }
 
@@ -106,7 +112,7 @@ public final class StackdriverStatsExporter {
     checkNotNull(credentials, "credentials");
     checkNotNull(projectId, "projectId");
     checkNotNull(exportInterval, "exportInterval");
-    createInternal(credentials, projectId, exportInterval, null);
+    createInternal(credentials, projectId, exportInterval, null, null);
   }
 
   /**
@@ -136,7 +142,7 @@ public final class StackdriverStatsExporter {
       throws IOException {
     checkNotNull(projectId, "projectId");
     checkNotNull(exportInterval, "exportInterval");
-    createInternal(null, projectId, exportInterval, null);
+    createInternal(null, projectId, exportInterval, null, null);
   }
 
   /**
@@ -159,6 +165,9 @@ public final class StackdriverStatsExporter {
    * cloud.google.com/monitoring/custom-metrics/creating-metrics#which-resource for a list of valid
    * {@code MonitoredResource}s.
    *
+   * <p>If {@code metricNamePrefix} of the configuration is not set, the exporter will use the
+   * default prefix "OpenCensus".
+   *
    * @param configuration the {@code StackdriverStatsConfiguration}.
    * @throws IllegalStateException if a Stackdriver exporter is already created.
    * @since 0.11.0
@@ -170,7 +179,8 @@ public final class StackdriverStatsExporter {
         configuration.getCredentials(),
         configuration.getProjectId(),
         configuration.getExportInterval(),
-        configuration.getMonitoredResource());
+        configuration.getMonitoredResource(),
+        configuration.getMetricNamePrefix());
   }
 
   /**
@@ -194,11 +204,13 @@ public final class StackdriverStatsExporter {
    *
    * <p>This method uses the default resource created from the environment variables.
    *
+   * <p>This method uses the default display name prefix "OpenCensus".
+   *
    * @throws IllegalStateException if a Stackdriver exporter is already created.
    * @since 0.11.0
    */
   public static void createAndRegister() throws IOException {
-    createInternal(null, null, null, null);
+    createInternal(null, null, null, null, null);
   }
 
   /**
@@ -225,7 +237,7 @@ public final class StackdriverStatsExporter {
   @Deprecated
   public static void createAndRegister(Duration exportInterval) throws IOException {
     checkNotNull(exportInterval, "exportInterval");
-    createInternal(null, null, exportInterval, null);
+    createInternal(null, null, exportInterval, null, null);
   }
 
   /**
@@ -254,7 +266,7 @@ public final class StackdriverStatsExporter {
     checkNotNull(projectId, "projectId");
     checkNotNull(exportInterval, "exportInterval");
     checkNotNull(monitoredResource, "monitoredResource");
-    createInternal(null, projectId, exportInterval, monitoredResource);
+    createInternal(null, projectId, exportInterval, monitoredResource, null);
   }
 
   /**
@@ -281,7 +293,7 @@ public final class StackdriverStatsExporter {
       Duration exportInterval, MonitoredResource monitoredResource) throws IOException {
     checkNotNull(exportInterval, "exportInterval");
     checkNotNull(monitoredResource, "monitoredResource");
-    createInternal(null, null, exportInterval, monitoredResource);
+    createInternal(null, null, exportInterval, monitoredResource, null);
   }
 
   // Use createInternal() (instead of constructor) to enforce singleton.
@@ -289,7 +301,8 @@ public final class StackdriverStatsExporter {
       @Nullable Credentials credentials,
       @Nullable String projectId,
       @Nullable Duration exportInterval,
-      @Nullable MonitoredResource monitoredResource)
+      @Nullable MonitoredResource monitoredResource,
+      @Nullable String metricNamePrefix)
       throws IOException {
     projectId = projectId == null ? ServiceOptions.getDefaultProjectId() : projectId;
     exportInterval = exportInterval == null ? DEFAULT_INTERVAL : exportInterval;
@@ -313,7 +326,8 @@ public final class StackdriverStatsExporter {
               metricServiceClient,
               exportInterval,
               Stats.getViewManager(),
-              monitoredResource);
+              monitoredResource,
+              metricNamePrefix);
       exporter.workerThread.start();
     }
   }