+++ CHANGELOG.md
@@ -7,7 +7,8 @@
   on recording exemplars.
 - Reduce the default limit on `Link`s per `Span` to 32 (was 128 before).
 - Add Spring support for `@Traced` annotation and java.sql.PreparedStatements 
-  tracing  
+  tracing
+- Allow custom prefix for Stackdriver metrics in `StackdriverStatsConfiguration`.
 
 ## 0.15.0 - 2018-06-20
 - Expose the factory methods of MonitoredResource.